move avatar upload into a separate function and move AvatarSettings to its own folder
